tests: newlib: thread_safety: Add internal lock tests
This commit adds the tests for the newlib retargetable locking
interface, as well as the tests for the internal lock functions that
are supposed to internally invoke the retargetable locking interface.

All of these tests must pass when the toolchain newlib is compiled with
the `retargetable-locking` and `multithread` options, which are
required to ensure that the newlib is thread-safe, enabled.

If the toolchain newlib is compiled with either of these options
disabled, this test will fail.

Signed-off-by: Stephanos Ioannidis <root@stephanos.io>